Add ordering on bill number for more consistency
Bill number is a string, and doesn't sort correctly across years. So we still sort via date introduced first, then use bill number to break ties (bills introduced on the same day), which wouldn't have issues sorting bill numbers across years (since it's all on the same day).

Code reuse is a bit fucked, because this spans both the front-end and the back-end. The scraping order should be the same as the order shown to the user.